requirements.txt
pyserial
chaco
pyyaml
pyqt5